embeddings-ada002/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==0.28.1
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.4.2